Move custom domain backfill to dated migration
